MNT: applied some suggestions
